Add files via upload
The above files are the codes for working of "solar seed fling machine". A simple arduino code is used to run the whole machine where one code represents the bluetooth module controlling dc motors and the other code is about servo motors rotation with a fixed speed and at a fixed angle.